Change exercise data structure
Previously was a key-value pair of answer: question. Will not work when
we want more than one question to have the same word as an answer.
Updated structure is now a list of object, with each object containing
an answer key-value pair and a question key-value pair.

No nesting to simplify existing front end logic, while allowing us to
easily get answer and question for each answer-question object.